Getting ready to work on first and second passes
Added a function to translate and print binary representation of integers.
Added relevant command's and label's members to reserved struct.
Created a struct codeLine to hold the data the assembly code lines. Added in main two arrays to hold each 10 codeLines of code & data. Added a function to initiate the arrays and the codeLine objects in it